Add capacitated telegraf to packageset
Adding a "capacitated" package means that is uses the same flox-floxpkgs
(and therefore nixpkgs) as the packageset instead of respecting the lock
of the package (in this case the lock in the github:etorreborre/telegraf
repo). This keeps the number of nixpkgs used by packages in a packageset
to a minimum, which prevents bloat. If this were not the case, it would
be possible, for example, to have many different versions of the same
libc in a single packageset.